* pokt/main: (51 commits)
  [RelayMiner] Supplier rate limiting (#895)
  [Tokenomics] refactor: `MintAllocationApplication` var usage to param usage  (#918)
  [Tokenomics] feat: add `mint_allocation_application` param to tokenomics module (#917)
  [Tokenomics] refactor: `MintAllocationSourceOwner` var usage to param usage (#916)
  [Tokenomics] feat: add `mint_allocation_source_owner` param to tokenomics module (#915)
  [Tokenomics] refactor: `MintAllocationSupplier` var usage to param usage (#914)
  [Tokenomics] feat: add `mint_allocation_supplier` param to tokenomics module (#913)
  [Tokenomics] refactor: `MintAllocationProposer` var usage to param usage (#912)
  [Tokenomics] feat: add `mint_allocation_proposer` param to tokenomics module (#911)
  [Tokenomics] refactor: `MintAllocationDao` var usage to param usage (#910)
  [Tokenomics] Add `mint_allocation_dao` tokenomics module param (#909)
  [quick-pr] change full-node.sh script links (#920)
  [Supplier] refactor: `NumSuppliersPerSession` var usage to param usage (#908)
  [Session Params] Add `num_suppliers_per_session` param to session module (#907)
  [Session Params] Add `MsgUpdateParam` to session module (#906)
  Updated some compiled proto files
  [TODOs] Update #2 to `TODO_BETA` to only reflect ACTUAL blockers (#900)
  [Workflow] Enforce `TODO_UPNEXT` comments to have an associated username (#896)
  [Observability] Foundation for load testing telemetry (#832)
  [Tokenomics] Implement Global Mint Reimbursement Request (#878)
  ...

+++ x/proof/keeper/proof_requirement_test.go
@@ -34,7 +34,7 @@ func TestKeeper_IsProofRequired(t *testing.T) {
 		numTrueSamples atomic.Int64
 	)
 
-	// TODO_BETA(@bryanchriswhite): This test is periodically flaky but theoretically shouldn't be.
+	// TODO_TECHDEBT(@bryanchriswhite): This test is periodically flaky but theoretically shouldn't be.
 	// What can we do to increase it's consistency without diving tolerance by 2?
 	sampleSize := poktrand.RequiredSampleSize(float64(probability), tolerance/2, confidence)
 
@@ -50,8 +50,8 @@ func TestKeeper_IsProofRequired(t *testing.T) {
 		}
 	}
 
-	expectedNumTrueSamples := float32(sampleSize) * probability
-	expectedNumFalseSamples := float32(sampleSize) * (1 - probability)
+	expectedNumTrueSamples := float64(sampleSize) * probability
+	expectedNumFalseSamples := float64(sampleSize) * (1 - probability)
 	toleranceSamples := tolerance * float64(sampleSize)
 	// Check that the number of samples for each outcome is within the expected range.
 	numFalseSamples := sampleSize - numTrueSamples.Load()
